+++ Cargo.toml
@@ -58,4 +58,4 @@ crate-type = ["cdylib","lib"]
 # ------------------------------
 
 [package.metadata.docs.rs]
-rustdoc-args = [ "--html-in-header", "./html/docs-header.html" ]
+rustdoc-args = [ "--html-in-header", "./html/rustdocs-header.html" ]
